added about and options screen and refactored MainMenuScreen (clean code)
